made syncvideo
Using the plyr media player I creadted a media player taht is synced across everyone on the website. I ran into a few problems because at the start I tried tolisten to the play and pause events which broke when two users were watching. The player got stuck n infinite loops and nothing worked. I switched the event listeners to ones that listened for clicks and keypresses that I controlled complelety myself. WHich worked out really well. It sent sockets that toled every user if that player was playing or  not and which timestamp the user who clicked was at, which made it synced up between every user.