Made usernames more secure
Removed all special characters from usernames when a user or a guest i registered. Made the error messages custom depending on what is wron and not just a standard error message. Got a tool for visual studio code where I can see and manage the data in the database, to make my development easier.